sqlalchemy/requirements.txt
sqlalchemy
pymysql
bcrypt
python-dotenv
snowflake-sqlalchemy